feat(api): refactor logic raf changes
Ref. metriport/metriport-internal#521
