+++ api/app/package.json
@@ -43,7 +43,6 @@
     "lodash": "^4.17.21",
     "mime-types": "^2.1.35",
     "nanoid": "^3.3.4",
-    "node-fetch": "^3.3.1",
     "oauth-signature": "^1.5.0",
     "pg": "^8.8.0",
     "pg-hstore": "^2.3.4",
@@ -67,7 +66,6 @@
     "@types/lodash": "^4.14.191",
     "@types/mime-types": "^2.1.1",
     "@types/node": "^18.11.12",
-    "@types/node-fetch": "^2.6.3",
     "@types/oauth-signature": "^1.5.0",
     "@types/sequelize": "^4.28.14",
     "@types/simple-oauth2": "^4.1.1",
