+++ api/app/src/external/commonwell/document/document-query.ts
@@ -12,15 +12,17 @@ import { makeCommonWellAPI, organizationQueryMeta } from "../api";
 import { getPatientData, PatientDataCommonwell } from "../patient-shared";
 import { downloadDocument } from "./document-download";
 import { upsertDocumentToFHIRServer } from "../../fhir/document/save-document-reference";
-import { toFHIR as toFHIRDocRef } from "./shared";
+import { toFHIR as toFHIRDocRef } from "../../fhir/document";
 import { DocumentWithFilename } from "./shared";
 import { Config } from "../../../shared/config";
 import { createS3FileName } from "../../../shared/external";
-import { debug } from "../../../shared/log";
 
 const s3client = new AWS.S3();
 
-export async function getDocuments({
+/**
+ * This is likely to be a long-running function
+ */
+export async function queryDocuments({
   patient,
   facilityId,
 }: {
@@ -28,26 +30,24 @@ export async function getDocuments({
   facilityId: string;
 }): Promise<void> {
   try {
-    const { organization, facility } = await getPatientData(patient, facilityId);
-
-    const cwDocuments = await internalGetDocuments({ patient, organization, facility });
-
     if (Config.isSandbox()) {
       return;
     }
 
-    console.log(debug(JSON.stringify(cwDocuments, null, 2)));
+    const { organization, facility } = await getPatientData(patient, facilityId);
 
-    const docsS3Refs = await dowloadAnduploadDocsToS3({
+    const cwDocuments = await internalGetDocuments({ patient, organization, facility });
+
+    const docsS3Refs = await dowloadAndUploadDocsToS3({
       patient,
       facilityId,
       documents: cwDocuments,
     });
 
-    docsS3Refs.forEach(async doc => {
+    for (const doc of docsS3Refs) {
       const FHIRDocRef = toFHIRDocRef(doc, organization, patient);
       await upsertDocumentToFHIRServer(FHIRDocRef);
-    });
+    }
   } catch (err) {
     console.log(`Error: `, err);
     capture.error(err, {
@@ -106,15 +106,18 @@ async function internalGetDocuments({
 
   const documents: Document[] = docs.entry
     ? docs.entry.flatMap(d =>
-        d.id && d.content && d.content.location
-          ? { id: d.id, content: { location: d.content.location, ...d.content } }
+        d.content?.masterIdentifier?.value && d.content && d.content.location
+          ? {
+              id: d.content.masterIdentifier.value,
+              content: { location: d.content.location, ...d.content },
+            }
           : []
       )
     : [];
   return documents;
 }
 
-async function dowloadAnduploadDocsToS3({
+async function dowloadAndUploadDocsToS3({
   patient,
   facilityId,
   documents,
@@ -125,14 +128,14 @@ async function dowloadAnduploadDocsToS3({
 }): Promise<DocumentWithFilename[]> {
   const uploadStream = (key: string) => {
     const pass = new stream.PassThrough();
-    const removePeriods = key.split(".").join("");
+    const base64key = Buffer.from(key).toString("base64");
 
     return {
       writeStream: pass,
       promise: s3client
         .upload({
           Bucket: Config.getMedicalDocumentsBucketName(),
-          Key: createS3FileName(patient.cxId, removePeriods),
+          Key: createS3FileName(patient.cxId, base64key),
           Body: pass,
         })
         .promise(),
@@ -141,46 +144,44 @@ async function dowloadAnduploadDocsToS3({
 
   const s3Refs = await Promise.allSettled(
     documents.map(async doc => {
-      if (doc.content?.masterIdentifier?.value) {
-        const { writeStream, promise } = uploadStream(doc.content?.masterIdentifier?.value);
+      try {
+        if (doc.content?.masterIdentifier?.value) {
+          const { writeStream, promise } = uploadStream(doc.content?.masterIdentifier?.value);
 
-        await downloadDocument({
-          cxId: patient.cxId,
-          patientId: patient.id,
-          facilityId: facilityId,
-          location: doc.content?.location ? doc.content.location : "",
-          stream: writeStream,
-        });
+          await downloadDocument({
+            cxId: patient.cxId,
+            patientId: patient.id,
+            facilityId: facilityId,
+            location: doc.content?.location ? doc.content.location : "",
+            stream: writeStream,
+          });
 
-        const data = await promise;
+          const data = await promise;
 
-        return {
-          ...doc,
-          content: {
-            ...doc.content,
-            location: data.Location,
+          return {
+            ...doc,
+            content: {
+              ...doc.content,
+              location: data.Location,
+            },
+            fileName: data.Key,
+          };
+        }
+
+        return undefined;
+      } catch (error) {
+        capture.error(error, {
+          extra: {
+            context: `s3.documentUpload`,
           },
-          fileName: data.Key,
-        };
+        });
       }
-
-      return undefined;
     })
   );
 
   const docsNewLocation: DocumentWithFilename[] = s3Refs.flatMap(ref =>
     ref.status === "fulfilled" && ref.value ? ref.value : []
   );
-
-  s3Refs.forEach(ref => {
-    if (ref.status === "rejected") {
-      capture.error(ref.reason, {
-        extra: {
-          context: `s3.documentUpload`,
-        },
-      });
-    }
-  });
 
   return docsNewLocation;
 }
