+++ api/app/src/external/fhir/document/remove-documents.ts
@@ -1,0 +1,9 @@
+// import { DocumentReference } from "@medplum/fhirtypes";
+// import { getDocuments } from "./get-documents";
+// import { api } from "../api";
+
+// export const removeDocuments = async (
+//   patientId: string
+// ): Promise<DocumentReference[] | undefined> => {
+//   return getDocuments(`patient=${patientId}&`);
+// };
