+++ api/app/src/external/fhir/organization/upsert-organization.ts
@@ -1,13 +1,6 @@
 import { Organization } from "@medplum/fhirtypes";
 import { api } from "../api";
-import { capture } from "../../../shared/notifications";
 
 export const upsertOrgToFHIRServer = async (organization: Organization) => {
-  try {
-    return await api.updateResource(organization);
-  } catch (err) {
-    capture.error(err, {
-      extra: { context: `fhir.add.organization` },
-    });
-  }
+  await api.updateResource(organization);
 };
