+++ api/app/src/external/fhir/patient/upsert-patient.ts
@@ -1,13 +1,6 @@
 import { Patient } from "@medplum/fhirtypes";
 import { api } from "../api";
-import { capture } from "../../../shared/notifications";
 
 export const upsertPatientToFHIRServer = async (patient: Patient) => {
-  try {
-    await api.updateResource(patient);
-  } catch (err) {
-    capture.error(err, {
-      extra: { context: `fhir.add.patient` },
-    });
-  }
+  await api.updateResource(patient);
 };
