--- api/app/src/models/medical/document-reference.ts
@@ -1,58 +1,0 @@
-import { DataTypes, Sequelize } from "sequelize";
-import {
-  DocumentReference,
-  ExternalDocumentReference,
-} from "../../domain/medical/document-reference";
-import { MedicalDataSource } from "../../external";
-import { BaseModel, ModelSetup } from "../_default";
-
-export type DocumentReferenceData = ExternalDocumentReference;
-
-export class DocumentReferenceModel
-  extends BaseModel<DocumentReferenceModel>
-  implements DocumentReference
-{
-  static NAME = "document_reference";
-  declare deletedAt?: Date;
-  declare cxId: string;
-  declare patientId: string;
-  declare source: MedicalDataSource;
-  declare externalId: string;
-  declare data: DocumentReferenceData;
-  declare raw?: unknown;
-
-  static setup: ModelSetup = (sequelize: Sequelize) => {
-    DocumentReferenceModel.init(
-      {
-        ...BaseModel.attributes(),
-        deletedAt: {
-          type: DataTypes.DATE(6),
-        },
-        cxId: {
-          type: DataTypes.UUID,
-        },
-        patientId: {
-          type: DataTypes.STRING,
-        },
-        source: {
-          type: DataTypes.STRING,
-        },
-        externalId: {
-          type: DataTypes.STRING,
-        },
-        data: {
-          type: DataTypes.JSONB,
-        },
-        raw: {
-          type: DataTypes.JSONB,
-        },
-      },
-      {
-        ...BaseModel.modelOptions(sequelize),
-        tableName: DocumentReferenceModel.NAME,
-        paranoid: true,
-        deletedAt: "deleted_at",
-      }
-    );
-  };
-}
