+++ api/app/src/routes/medical/organization.ts
@@ -5,7 +5,7 @@ import {
   createOrganization,
   OrganizationCreateCmd,
 } from "../../command/medical/organization/create-organization";
-import { toFHIR } from "../../external/commonwell/organization";
+import { toFHIR } from "../../external/fhir/organization";
 import { upsertOrgToFHIRServer } from "../../external/fhir/organization/upsert-organization";
 import { getOrganization } from "../../command/medical/organization/get-organization";
 import {
